src/data_acquisition/sinca_scraper.py
"""
SINCA Data Scraper
Downloads PM2.5 ground-truth data from Chilean air quality monitoring network
"""

import requests
import pandas as pd
from bs4 import BeautifulSoup
from datetime import datetime, timedelta
from pathlib import Path
import logging
import time
from typing import List, Dict
import yaml

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


class SINCADataScraper:
    """
    Scraper for SINCA (Sistema de Información Nacional de Calidad del Aire) data.
    """

    def __init__(self, config_path: str = "config/config.yaml"):
        """
        Initialize SINCA scraper.

        Args:
            config_path: Path to configuration file
        """
        self.config = self._load_config(config_path)
        self.base_url = self.config['sinca']['base_url']
        self.stations = self.config['sinca']['stations']
        self.pollutants = self.config['sinca']['pollutants']

    def _load_config(self, config_path: str) -> Dict:
        """Load configuration from YAML file."""
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        return config

    def download_station_data(
        self,
        station: str,
        pollutant: str,
        start_date: str,
        end_date: str
    ) -> pd.DataFrame:
        """
        Download data for a specific station and pollutant.

        Args:
            station: Station name
            pollutant: Pollutant type (PM25, PM10, etc.)
            start_date: Start date (YYYY-MM-DD)
            end_date: End date (YYYY-MM-DD)

        Returns:
            DataFrame with time series data
        """
        logger.info(f"Downloading {pollutant} data for {station}: {start_date} to {end_date}")

        # NOTE: This is a template - actual implementation depends on SINCA API/website structure
        # You may need to:
        # 1. Use Selenium for JavaScript-rendered pages
        # 2. Find the actual API endpoints
        # 3. Handle authentication if required

        # Placeholder implementation
        # TODO: Replace with actual SINCA data access method

        try:
            # Example: If SINCA has a direct download API
            params = {
                'station': station,
                'pollutant': pollutant,
                'start_date': start_date,
                'end_date': end_date
            }

            # This is a placeholder URL - replace with actual SINCA endpoint
            # response = requests.get(f"{self.base_url}/data", params=params)
            # response.raise_for_status()

            # For now, create empty DataFrame with expected structure
            logger.warning("Using placeholder data - implement actual SINCA scraping!")

            date_range = pd.date_range(start=start_date, end=end_date, freq='H')
            df = pd.DataFrame({
                'datetime': date_range,
                'station': station,
                'pollutant': pollutant,
                'value': None  # Replace with actual values
            })

            return df

        except Exception as e:
            logger.error(f"Error downloading data for {station}: {e}")
            return pd.DataFrame()

    def download_manual_csv(
        self,
        csv_path: str,
        output_dir: str = "data/raw/sinca"
    ) -> pd.DataFrame:
        """
        Load SINCA data from manually downloaded CSV files.

        This is a workaround if automatic scraping is not possible.
        Users can manually download data from SINCA website and place in data/external/

        Args:
            csv_path: Path to CSV file
            output_dir: Output directory

        Returns:
            Cleaned DataFrame
        """
        logger.info(f"Loading SINCA data from CSV: {csv_path}")

        try:
            # Read CSV
            df = pd.read_csv(csv_path)

            # Expected columns: datetime, station, pollutant, value
            # Adjust based on actual SINCA CSV format

            # Parse datetime
            if 'datetime' in df.columns:
                df['datetime'] = pd.to_datetime(df['datetime'])

            # Clean data
            df = df.dropna(subset=['value'])
            df = df[df['value'] >= 0]  # Remove negative values

            # Save cleaned version
            Path(output_dir).mkdir(parents=True, exist_ok=True)
            output_file = Path(output_dir) / f"sinca_cleaned_{Path(csv_path).stem}.csv"
            df.to_csv(output_file, index=False)
            logger.info(f"Saved cleaned data to {output_file}")

            return df

        except Exception as e:
            logger.error(f"Error loading CSV: {e}")
            return pd.DataFrame()

    def download_all_stations(
        self,
        start_date: str,
        end_date: str,
        pollutant: str = "PM25",
        output_dir: str = "data/raw/sinca"
    ):
        """
        Download data for all stations in Santiago.

        Args:
            start_date: Start date (YYYY-MM-DD)
            end_date: End date (YYYY-MM-DD)
            pollutant: Pollutant to download
            output_dir: Output directory
        """
        logger.info(f"Downloading {pollutant} for all stations: {start_date} to {end_date}")

        Path(output_dir).mkdir(parents=True, exist_ok=True)

        all_data = []

        for station in self.stations:
            try:
                df = self.download_station_data(station, pollutant, start_date, end_date)

                if not df.empty:
                    all_data.append(df)

                # Rate limiting
                time.sleep(2)

            except Exception as e:
                logger.error(f"Error with station {station}: {e}")
                continue

        # Combine all data
        if all_data:
            combined_df = pd.concat(all_data, ignore_index=True)

            # Save
            output_file = Path(output_dir) / f"sinca_{pollutant}_{start_date}_{end_date}.csv"
            combined_df.to_csv(output_file, index=False)
            logger.info(f"Saved combined data to {output_file}")
            logger.info(f"Total records: {len(combined_df)}")

            return combined_df
        else:
            logger.warning("No data downloaded!")
            return pd.DataFrame()

    def get_station_metadata(self) -> pd.DataFrame:
        """
        Get metadata for SINCA stations (location, elevation, etc.).

        Returns:
            DataFrame with station metadata
        """
        # Hardcoded metadata for Santiago stations
        # TODO: Scrape this from SINCA website or use official metadata

        metadata = {
            'station': [
                'Cerrillos', 'Cerro Navia', 'El Bosque', 'Independencia',
                'La Florida', 'Las Condes', 'Pudahuel', 'Puente Alto'
            ],
            'lat': [
                -33.50, -33.42, -33.56, -33.41,
                -33.52, -33.40, -33.44, -33.61
            ],
            'lon': [
                -70.71, -70.74, -70.69, -70.66,
                -70.58, -70.58, -70.77, -70.58
            ],
            'elevation': [
                522, 485, 580, 540,
                600, 700, 475, 640
            ]
        }

        df = pd.DataFrame(metadata)

        # Save
        output_file = "data/external/sinca_stations.csv"
        Path("data/external").mkdir(parents=True, exist_ok=True)
        df.to_csv(output_file, index=False)
        logger.info(f"Saved station metadata to {output_file}")

        return df


def main():
    """Main execution function."""
    import argparse

    parser = argparse.ArgumentParser(description="Download SINCA data")
    parser.add_argument("--config", default="config/config.yaml", help="Config file")
    parser.add_argument("--start-date", required=True, help="Start date (YYYY-MM-DD)")
    parser.add_argument("--end-date", required=True, help="End date (YYYY-MM-DD)")
    parser.add_argument("--pollutant", default="PM25", help="Pollutant type")
    parser.add_argument("--manual-csv", help="Path to manually downloaded CSV")

    args = parser.parse_args()

    scraper = SINCADataScraper(config_path=args.config)

    if args.manual_csv:
        # Load from manual CSV
        scraper.download_manual_csv(args.manual_csv)
    else:
        # Automatic download (needs implementation)
        scraper.download_all_stations(
            start_date=args.start_date,
            end_date=args.end_date,
            pollutant=args.pollutant
        )

    # Also download station metadata
    scraper.get_station_metadata()


if __name__ == "__main__":
    main()
