requirements.txt
# PM2.5 Santiago - Python Dependencies
# Updated: 2025-11-10

# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Google Earth Engine
earthengine-api>=0.1.380
geemap>=0.30.0

# Geospatial libraries
geopandas>=0.14.0
rasterio>=1.3.0
fiona>=1.9.0
shapely>=2.0.0
pyproj>=3.5.0
rtree>=1.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning (optional, for future work)
# tensorflow>=2.13.0
# torch>=2.0.0

# Hyperparameter tuning
optuna>=3.3.0
hyperopt>=0.2.7

# Model interpretation
shap>=0.42.0
lime>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
folium>=0.14.0
contextily>=1.3.0

# Web scraping (for SINCA data)
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
lxml>=4.9.0

# Data handling
h5py>=3.9.0
netCDF4>=1.6.0
xarray>=2023.5.0

# Progress bars and utilities
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
ipykernel>=6.25.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Statistical analysis
statsmodels>=0.14.0
pingouin>=0.5.3

# Time series
prophet>=1.1.4

# Quality assurance
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Export and reporting
openpyxl>=3.1.0
fpdf2>=2.7.0
reportlab>=4.0.0
